+++ crack-core/src/commands/music/play.rs
@@ -1,7 +1,8 @@
 use self::serenity::builder::CreateEmbed;
 use crate::{
-    commands::{aux_metadata_to_db_structures, skip::force_skip_top_track},
+    commands::skip::force_skip_top_track,
     connection::get_voice_channel_for_user,
+    db::aux_metadata_to_db_structures,
     errors::{verify, CrackedError},
     guild::settings::{GuildSettings, DEFAULT_PREMIUM},
     handlers::{track_end::update_queue_messages, IdleHandler, TrackEndHandler},
