+++ crack-core/src/commands/playlist/create_playlist.rs
@@ -5,8 +5,8 @@ use crate::{
 
 /// Creates a playlist
 #[cfg(not(tarpaulin_include))]
-#[poise::command(prefix_command, slash_command)]
-pub async fn create(ctx: Context<'_>, name: String) -> Result<(), Error> {
+#[poise::command(prefix_command, slash_command, rename = "create")]
+pub async fn create_playlist(ctx: Context<'_>, name: String) -> Result<(), Error> {
     // Assuming you have a way to fetch the user_id of the command issuer
     let user_id = ctx.author().id.get() as i64;
