+++ crack-core/src/commands/playlist/mod.rs
@@ -2,9 +2,9 @@ pub mod add_to_playlist;
 pub mod create_playlist;
 pub mod delete_playlist;
 
-pub use add_to_playlist::*;
-pub use create_playlist::*;
-pub use delete_playlist::*;
+pub use add_to_playlist::add_to_playlist as add;
+pub use create_playlist::create_playlist as create;
+pub use delete_playlist::delete_playlist as delete;
 
 use crate::{Context, Error};
